functions: split updateSummaries, extracting helper
